requirements.txt
django
django-allauth
twython
pyyaml
pymongo
pygeocoder
requests
pillow
mailchimp

dj-stripe
django-floppyforms
urllib3
mandrill
pyDNS
validate_email
twilio